fixup: fix for comment
