+++ .github/workflows/check-generated.yml
@@ -26,7 +26,7 @@ jobs:
 
       - name: Check generated error docs
         run: |
-          make error-doc
+          make error-doc-gen
           if ! git diff --stat --exit-code ; then
             echo ">> ERROR:"
             echo ">>"
