--- tools/error_doc/generator/module_directory_walker.go
@@ -1,22 +1,0 @@
-package generator
-
-import (
-	"os"
-	"path/filepath"
-)
-
-func (edg *ErrorDocumentGenerator) listUpErrorsGoFiles(startPath, errorsFileName string) error {
-	err := filepath.Walk(startPath, func(path string, info os.FileInfo, err error) error {
-		if err != nil {
-			return err
-		}
-		if !info.IsDir() && info.Name() == errorsFileName {
-			edg.errorsFiles = append(edg.errorsFiles, path)
-		}
-		return nil
-	})
-	if err != nil {
-		return err
-	}
-	return nil
-}
